+++ Game/src/Game/Attibutes/Bonus.cpp
@@ -2,7 +2,7 @@
 
 #include <Game/Attibutes/Bonus.hpp>
 
-Bonus::Bonus(double baseValue) : _value(baseValue)
+Bonus::Bonus(int baseValue) : _value(baseValue)
 {
 }
 
@@ -10,7 +10,7 @@ Bonus::~Bonus()
 {
 }
 
-double Bonus::getValue() const
+int Bonus::getValue() const
 {
     return this->_value;
 }
